+++ Pandora/Paradigm/Primary.hs
@@ -1,4 +1,5 @@
 module Pandora.Paradigm.Primary (module Exports) where
 
-import Pandora.Paradigm.Primary.Functor as Exports
 import Pandora.Paradigm.Primary.Transformer as Exports
+import Pandora.Paradigm.Primary.Functor as Exports
+import Pandora.Paradigm.Primary.Object as Exports