+++ Pandora/Paradigm/Primary/Functor.hs
@@ -1,4 +1,4 @@
-module Pandora.Paradigm.Primary.Functor where
+module Pandora.Paradigm.Primary.Functor (module Exports, note, hush, left, right, this, that, here, there) where
 
 import Pandora.Paradigm.Primary.Functor.Fix as Exports
 import Pandora.Paradigm.Primary.Functor.Predicate as Exports