+++ Pandora/Paradigm/Primary/Object.hs
@@ -1,0 +1,3 @@
+module Pandora.Paradigm.Primary.Object (module Exports) where
+
+import Pandora.Paradigm.Primary.Object.Boolean as Exports