+++ CHANGELOG.md
@@ -224,3 +224,4 @@
 * Rename `Variation` to `These`
 * Define two experimental methods in `Bindable` class - `$>>=` and `>>=$`
 * Define `here` and `there` methods from `Wedge` to `Maybe`
+* Move `Boolean` definition to its own module